pyproject.toml
[project]
name = "agenticbloodbank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-core>=0.7.5",
    "autogen-ext>=0.7.5",
    "fastapi>=0.122.0",
    "firebase-admin>=7.1.0",
    "genai>=2.1.0",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
